Update mobile ci and separate mobile and mock tests
