+++ apps/ledger-live-mobile/scripts/e2e-ci.mjs
@@ -2,7 +2,7 @@
 import { basename } from "path";
 
 let platform, test, build, bundle;
-let speculos = "";
+let testType = "mock";
 let cache = true;
 let shard = "";
 let target = "release";
@@ -18,7 +18,7 @@ const usage = (exitCode = 1) => {
   console.log(
     `Usage: ${basename(
       __filename,
-    )} -p --platform <ios|android> [-h --help]  [-t --test] [-b --build] [--bundle] [--cache | --no-cache] [--speculos] [--shard] [--production]`,
+    )} -p --platform <ios|android> [-h --help]  [-t --test] [-b --build] [--bundle] [--cache | --no-cache] [--testType] [--shard] [--production]`,
   );
   process.exit(exitCode);
 };
@@ -46,7 +46,7 @@ const bundle_ios_with_cache = async () => {
 };
 
 const test_ios = async () => {
-  await $`pnpm mobile e2e:test${speculos} \
+  await $`pnpm mobile ${testType}:test\
     -c ios.sim.release \
     --loglevel error \
     --record-logs failing \
@@ -64,7 +64,7 @@ const build_android = async () => {
 };
 
 const test_android = async () => {
-  await $`pnpm mobile e2e:test${speculos} \\
+  await $`pnpm mobile ${testType}:test \\
     -c android.emu.${target} \\
     --loglevel error \\
     --record-logs failing \\
@@ -121,8 +121,8 @@ for (const argName in argv) {
       break;
     case "_":
       break;
-    case "speculos":
-      speculos = ":speculos";
+    case "e2e":
+      testType = "e2e";
       break;
     case "shard":
       shard = argv[argName];
