+++ apps/ledger-live-mobile/src/screens/Swap/LiveApp/customHandlers/index.ts
@@ -6,7 +6,7 @@ import { getFee } from "./getFee";
 import { getTransactionByHash } from "./getTransactionByHash";
 import { saveSwapToHistory } from "./saveSwapToHistory";
 import Config from "react-native-config";
-import { sendSwapLiveAppReady } from "../../../../../e2e/bridge/client";
+import { sendSwapLiveAppReady } from "../../../../../../../e2e/mobile/bridge/client";
 
 export type NavigationType = Omit<NavigationProp<ReactNavigation.RootParamList>, "getState"> & {
   getState(): NavigationState | undefined;
