+++ package.json
@@ -106,6 +106,7 @@
     "domain": "pnpm --filter domain-service",
     "doc": "pnpm --filter docs",
     "mobile": "pnpm --filter live-mobile",
+    "e2e:mobile": "pnpm --filter ledger-live-mobile-e2e-tests",
     "common": "pnpm --filter live-common",
     "device-core": "pnpm --filter device-core",
     "device-react": "pnpm --filter device-react",
